+++ Releasenotes.md
@@ -163,6 +163,9 @@ Feature Release 3.3.0
 Major Release 4.0.0
 ===================
 
+Date: 16.12.2016
+Tag: 4.0.0
+
 * Breaking Change: Token can now contain annotations
 * Deprecation: Cached and Compiled Templates are deprecated and are no longer actively maintained - will be removed compeletely in next major version
 * Breaking Change / Deprecation: Disabled enabledInterpretedTemplateCache and deprecated all uses in Engine
@@ -170,3 +173,8 @@ Major Release 4.0.0
 * New Feature: Added journaling error handler (JournalingErrorHandler) and OutputAppender abstraction to embed error messages directly in output using ErrorReportingOutputAppender  (see ErrorReportingTest)
 * New Feature: When iterating over maps you can now choose between values, keys, and entries (map._values, map._keys, map.entries) 
 * New Feature: Allow renderers to be applied before comparing in if statements (see BooleanIfRendererTest)
+
+Release 4.1
+===========
+
+Date: so far only snapshot